+++ src/Model/Forward.java
@@ -5,11 +5,11 @@ import java.util.List;
 public class Forward extends Node {
 
     @Override
-    public int interpret() {
+    public double interpret() {
         Turtle turtle = getTurtle();
         List<Node> children = getChildren();
 
-        int dist = children.get(0).interpret();
+        double dist = children.get(0).interpret();
         int dir = turtle.getDirection();
 
         if (turtle != null) {
