+++ src/Model/Repeat.java
@@ -7,10 +7,10 @@ public class Repeat extends Node {
 	// child 0 is number of repeats
 	// child 1 is list to repeat
 	@Override
-	public int interpret() {
+	public double interpret() {
 		List<Node> children = getChildren();
-		int numIter = children.get(0).interpret();
-		int ret = 0;
+		double numIter = children.get(0).interpret();
+		double ret = 0;
 		
 		if (Controller.Controller.repCount == null) {
 			Controller.Controller.repCount = 0;
